Fixed Email Service and Created News Service
Updated EmailClass to fix an issue the filename creates in some IDE's

Created NewsPull, pulls up to date financial news from bloomberg-market-and-financial-news.p.rapidapi.com API

Created NewsReport, generates a PDF which can be attached as  an attachment to email to users

These files reference a file named confidential, this file is used to hide personal information.  This need will be eliminated once formatting a temp file on the server and setting up a gmail account for the website.